GPIO: Use enums for configuration
In the past, macros were used to setup GPIO pins, however this is prone
to errors and misconfiguration. Changing these to enums will ensure the
correct values are passed as arguments.

This commit also improves GPIO documentation, including references back
to the original libopencm3 docs.

+++ src/core/stm32f4/gpio.cpp
@@ -24,8 +24,8 @@ GPIO::GPIO(Pinout pin, Mode mode, Pullup pullup, Speed speed, DriverType driver)
 }
 
 void GPIO::Init(const Mode mode, const Pullup pullup, const Speed speed, const DriverType driver) const {
-  gpio_mode_setup(pin_.first, mode, pullup, pin_.second);
-  gpio_set_output_options(pin_.first, driver, speed, pin_.second);
+  gpio_mode_setup(pin_.first, static_cast<uint8_t>(mode), static_cast<uint8_t>(pullup), pin_.second);
+  gpio_set_output_options(pin_.first, static_cast<uint8_t>(driver), static_cast<uint8_t>(speed), pin_.second);
 }
 
 void GPIO::InitRcc(const Port port) const {
